sp_firebase/requirements.txt
Flask==0.12.2
Flask-User
google-api-python-client
google-auth-oauthlib
google-auth
httplib2
oauth2client
google-auth-httplib2
firebase-admin